requirements-client.txt
grpcio
numpy
opencv-python
torch
sounddevice
pynput
scipy
speechrecognition
openai-whisper

